Alex's first commit
Created Initializer, which is a singleton object that starts up the
system by importing the text files, instantiating the singleton
University, instantiating the singleton LoginManager (the
LoginManager's constructor calls LoginManager.displayLoginScreen); it
also includes a method to save all files and shutdown (if it receives
a message from the Login Manager that the actor has shutdown. Also
created the University object, which owns all the managers, and
vectors for students and courses.