Created Initializer.importPasswords() via Observer Pattern
Created class enum SystemStatus which indicates current system status.
Possible values are STARTING_UP (Initializer is in control of system,
importing, populating data objects, and running data validation tests;
RUNNING (University is in control of the system and available for users
to interact with system; SHUTTING_DOWN (the ShutDownManager is in
control of the system as it is saving data to text files).

Created an Observer pattern with the Subject & Observer interfaces.
Currently, the University implements the Subject interface. During
instantiation, it registers all the Managers as observers and implements
methods to notify the managers of changes to the University's
systemStatus.

The PasswordManager was updated to remove it's instantiation of the
University Object and added the method importPasswords(), which receives
an ArrayList<Password> object from Initializer and copies it to the
passwords container IF AND ONLY IF SystemStatus.STARTING_UP is current
state.